+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-
+* Added *-a , --validate-all* option in *validate* to run all validation on all files.
 
 #### 0.5.2
 * Added *-c, --command* option in *generate-docs* to generate a specific command from an integration.